requirements.txt
asgiref==3.11.0
binary==1.0.2
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
psycopg2==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
razorpay==2.0.0
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
